+++ timmons-group/src/config-form/components/AnyField/AnyField.jsx
@@ -10,14 +10,14 @@ import {
 } from '@mui/material';
 import { DatePicker } from '@mui/x-date-pickers';
 
-import RadioOptions from '../../components/RadioOptions';
-import Typeahead from '../../components/Typeahead';
-import FormErrorMessage from '../../components/FormErrorMessage';
-import AnyFieldLabel from '../../components/AnyFieldLabel';
+import RadioOptions from '../RadioOptions';
+import Typeahead from '../Typeahead';
+import FormErrorMessage from '../FormErrorMessage';
+import AnyFieldLabel from '../AnyFieldLabel';
 
-import { FIELD_TYPES } from '../../constants';
-import { isObject } from '../../helpers';
-import { dateStringNormalizer, isEmpty } from '../../helpers/helpers';
+import { FIELD_TYPES } from '../../../shared-react-components/constants';
+import { isObject } from '../../../shared-react-components/helpers';
+import { dateStringNormalizer, isEmpty } from '../../../shared-react-components/helpers/helpers';
 
 const makeFilter = (checkedId) => {
   const checkId = isObject(checkedId) ? checkedId.id : checkedId;